Dad 144 python link in header returns 404 (#359)
* copied navbar-version-selector.html to docs site partials.

<!--- Removed the {{.url}} so we could use this to link to sdk docs. Hugo was prepending the docsite url to the sdk url --->
<!-- <a class="dropdown-item" href="{{ .url }}{{ $path }}">{{ .version }}</a> --->

This is the line being used:
<a class="dropdown-item" href="{{ $path }}">{{ .version }}</a>

Note: This can’t be verified outside of prod. Hugo doesn’t build the sdk url the same way in local mode. To verify before merging: Run, “Hugo” to build the html locally. Open index.html. Verify that the SDK link on ~L.69 is correct. It should be: /https:/python.viam.dev/

* Removed {{ .url }} from line 10. That lets hugo use the right URL in  {{ .path }}