Updated 404 page.
You may not be able to review from the PR preview. Once our PR preview site is open, add 404.html to the end if the URL and press enter. If that doesn't work, you'll need to look in the branch. Hugo _won't_ serve this 404 page. It relies on outside servers to manage that.