Update docs for "APP-3453 Do not bill for cloud slam compute"
Since cloud slam compute is quite expensive under our current pricing model, but it's a large effort to add a separate billing sku for cloud slam compute, for the moment we just stopped charging for it under APP-3453.

I wanted to suggest docs changes to reflect this, and marked APP-3453 as "Docs changes needed", but couldn't figure out if a DOCS ticket was automatically generated. So, I'm just putting up this PR, and adding it to the "Docs changes needed" description on APP-3453.